+++ src/freenutritionfacts.nim
@@ -1,13 +1,13 @@
 #[
   Created at: 09/24/2021 20:57:44 Friday
-  Modified at: 09/24/2021 11:26:16 PM Friday
+  Modified at: 10/15/2021 01:30:53 PM Friday
 
-        Copyright (C) 2021 Thiago Navarro
+        Copyright (C) 2021 Thiago Ferreira
   See file "license" for details about copyright
 ]#
 
-import freenutritionfacts/props
-export props
+import freenutritionfacts/foods
+export foods
 
 import freenutritionfacts/categories
 export categories