+++ tests/test1.nim
@@ -1,8 +1,8 @@
 #[
   Created at: 09/24/2021 20:35:25 Friday
-  Modified at: 09/24/2021 08:35:25 PM Friday
+  Modified at: 10/15/2021 01:30:43 PM Friday
 
-        Copyright (C) 2021 Thiago Navarro
+        Copyright (C) 2021 Thiago Ferreira
   See file "license" for details about copyright
 ]#