+++ changelog.md
@@ -1,14 +1,22 @@
 <!--
   Created at: 09/24/2021 21:05:03 Friday
-  Modified at: 09/24/2021 11:28:05 PM Friday
+  Modified at: 10/15/2021 04:49:28 PM Friday
 
-        Copyright (C) 2021 Thiago Navarro
+        Copyright (C) 2021 Thiago Ferreira
   See file "license" for details about copyright
 -->
 
 # Changelog
 
 Changes
+
+## Version 1.1.0 (10/15/2021 13:28:23 Friday)
+
+- Reduced Nim minimum version to `1.0.0`
+- Renamed module `freenutritionfacts.props` to `freenutritionfacts.foods`
+- cleanups
+
+---
 
 ## Version 1.0.0 (09/26/2021)
 
@@ -17,7 +25,6 @@ Changes
 - Reorganized project
 
 ---
-
 
 ## Version 0.3.0 (09/25/2021)